+++ src/app.controller.ts
@@ -3,7 +3,7 @@ import { Controller, Get } from '@nestjs/common';
 @Controller()
 export class AppController {
   @Get()
-  public hello(): string {
+  public getHello(): string {
     return 'Hello World!';
   }
 }
