+++ src/app.module.ts
@@ -5,20 +5,23 @@ import { ConfigModule } from '@nestjs/config';
 import { AppController } from './app.controller';
 import { EtsController } from './common/api-helper/ets/ets.controller';
 import { EtsModule } from './common/api-helper/ets/ets.module';
-import { ScraperModule } from './common/api-helper/ets/scraper/scraper.module';
 import { FileUtil } from './common/utils/pdf/fileUtil';
 import { PdfController } from './common/website-helper/pdf/pdf.controller';
 import { HoraireCoursService } from './common/website-helper/pdf/pdf-parser/horaire/horaire-cours.service';
 import { PlanificationCoursService } from './common/website-helper/pdf/pdf-parser/planification/planification-cours.service';
 import config from './config/configuration';
+import { CourseController } from './course/course.controller';
 import { CourseModule } from './course/course.module';
+import { CourseInstanceController } from './course-instance/course-instance.controller';
 import { CourseInstanceModule } from './course-instance/course-instance.module';
 import { CoursePrerequisiteController } from './course-prerequisite/course-prerequisite.controller';
 import { CoursePrerequisiteModule } from './course-prerequisite/course-prerequisite.module';
 import { CoursePrerequisiteService } from './course-prerequisite/course-prerequisite.service';
 import { PrismaModule } from './prisma/prisma.module';
 import { PrismaService } from './prisma/prisma.service';
+import { ProgramController } from './program/program.controller';
 import { ProgramModule } from './program/program.module';
+import { SessionController } from './session/session.controller';
 import { SessionModule } from './session/session.module';
 
 @Module({
@@ -34,7 +37,6 @@ import { SessionModule } from './session/session.module';
     CourseInstanceModule,
     SessionModule,
     CoursePrerequisiteModule,
-    ScraperModule,
     EtsModule,
     PrismaModule,
   ],
@@ -49,7 +51,11 @@ import { SessionModule } from './session/session.module';
     AppController,
     PdfController,
     EtsController,
+    CourseController,
+    CourseInstanceController,
     CoursePrerequisiteController,
+    ProgramController,
+    SessionController,
   ],
 })
 export class AppModule {}
