+++ src/course/course.controller.ts
@@ -7,12 +7,12 @@ export class CourseController {
   constructor(private readonly courseService: CourseService) {}
 
   @Get(':id')
-  public course(@Param('id') id: string) {
-    return this.courseService.course({ id });
+  public getCourse(@Param('id') id: string) {
+    return this.courseService.getCourse({ id });
   }
 
   @Get()
-  public courses() {
-    return this.courseService.courses();
+  public getAllCourses() {
+    return this.courseService.getAllCourses();
   }
 }
