+++ test/course-prerequisite/course-prerequisite.controller.spec.ts
@@ -1,13 +1,15 @@
 import { Test, TestingModule } from '@nestjs/testing';
 
+import { CoursePrerequisiteController } from '../../src/course-prerequisite/course-prerequisite.controller';
+import { CoursePrerequisiteService } from '../../src/course-prerequisite/course-prerequisite.service';
 import { PrismaService } from '../../src/prisma/prisma.service';
-import { CoursePrerequisiteService } from './course-prerequisite.service';
 
-describe('CoursePrerequisiteService', () => {
-  let service: CoursePrerequisiteService;
+describe('CoursePrerequisiteController', () => {
+  let controller: CoursePrerequisiteController;
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
+      controllers: [CoursePrerequisiteController],
       providers: [
         CoursePrerequisiteService,
         {
@@ -16,19 +18,18 @@ describe('CoursePrerequisiteService', () => {
             coursePrerequisite: {
               findMany: jest.fn(),
               findUnique: jest.fn(),
-              create: jest.fn(),
-              update: jest.fn(),
-              delete: jest.fn(),
             },
           },
         },
       ],
     }).compile();
 
-    service = module.get<CoursePrerequisiteService>(CoursePrerequisiteService);
+    controller = module.get<CoursePrerequisiteController>(
+      CoursePrerequisiteController,
+    );
   });
 
   it('should be defined', () => {
-    expect(service).toBeDefined();
+    expect(controller).toBeDefined();
   });
 });
